Reconciles the cache-group-router rebuild with main's parallel work:

- glm-5.3-flash (#1259): the GLM53 DSA+KDA hybrid family joins the
  _AttnSideProfile facts (is_dsa_kda -> glm53_flash cache family, 'dsa'
  full-attention backend); the DSA leaf gains the KPoolRuntime surface
  (require_kpool_runtime, kpool_prefill/decode_page_table, per-forward
  reset) and the kv_seq_lens plumbing for pooled top-k; the router
  delegates the KPool accessors to its sole leaf. glm53_flash model files
  migrate to the Phase-4 write-location contract (no out_cache_loc
  threading; per-mode write_locations fetches around the prefill/decode
  split, cache_num_tokens narrowing preserved). A KPool-configured DSA
  leaf re-enables the prefill graph per instance (GLM53 handles padded
  replay explicitly).
- deepseek_v3: main's cache_num_tokens narrowing (partial-row KV writes
  for GLM's sparse dispatch) lands on the explicit-span form.
- MLAPrefillMetadata carries CPU tensors (extend_prefix_lens_cpu /
  extend_seq_lens_cpu) per main; registry adopts dataclasses.replace for
  named-backend spec routing.
- MiniMax M2 removal (#1365) and the V4 NextN rename (#1364) apply
  cleanly; kimi-k3 fused projections (#1358) auto-merge.
- Tests: stale hybrid_linear_attn/hybrid_kda module paths repointed to
  mamba/kda; the DSA dense-routing test pins the leaf constructor path;
  main's two get_capture_warmup_seq_len tests dropped (the helper does
  not exist anywhere on main — lost in the glm53 squash).

+++ python/tokenspeed/runtime/layers/attention/backends/mamba.py
@@ -120,8 +120,8 @@ def _compute_state_block_index_plan(
     seq_lens_before: torch.Tensor,
     seq_lens_after: torch.Tensor,
 ) -> _StateBlockIndexPlan:
-    before = seq_lens_before.to(torch.int64)
-    after = seq_lens_after.to(torch.int64)
+    before = seq_lens_before
+    after = seq_lens_after
     in_slots = torch.div(
         before - 1, checkpoint_granularity, rounding_mode="floor"
     ).clamp_(min=0)
@@ -710,7 +710,7 @@ class MambaAttnBackend(AttentionBackend):
         state_in_by_group = self.forward_metadata.state_in_blocks_by_group
         sin_stack = torch.stack(
             [state_in_by_group[group_id][:bs] for group_id in self._state_groups()]
-        ).to(torch.int64)
+        )
         src_rows = sin_stack.index_select(0, tables["group_sel"]).reshape(-1)
         copy_state_rows(
             tables["conv_comp"],
